test: fix flakiness of stringbytes-external
The tests used to rely on precise timing of when a JavaScript object
would be garbage collected to ensure that there is enough memory
available on the system. Switch the test to use a malloc/free pair
instead.

Ref: https://github.com/nodejs/node/pull/5945
Ref: https://github.com/nodejs/node/pull/6039
Ref: https://github.com/nodejs/node/pull/6073
