+++ test/addons/stringbytes-external-exceed-max/test-stringbytes-external-exceed-max-by-1-ascii.js
@@ -1,7 +1,7 @@
 'use strict';
-// Flags: --expose-gc
 
-const common = require('../common');
+const common = require('../../common');
+const binding = require('./build/Release/binding');
 const assert = require('assert');
 
 const skipMessage =
@@ -10,7 +10,6 @@ if (!common.enoughTestMem) {
   console.log(skipMessage);
   return;
 }
-assert(typeof gc === 'function', 'Run this test with --expose-gc');
 
 // v8 fails silently if string length > v8::String::kMaxLength
 // v8::String::kMaxLength defined in v8.h
@@ -18,9 +17,6 @@ const kStringMaxLength = process.binding('buffer').kStringMaxLength;
 
 try {
   var buf = new Buffer(kStringMaxLength + 1);
-  // Try to allocate memory first then force gc so future allocations succeed.
-  new Buffer(2 * kStringMaxLength);
-  gc();
 } catch (e) {
   // If the exception is not due to memory confinement then rethrow it.
   if (e.message !== 'Invalid array buffer length') throw (e);
@@ -28,14 +24,12 @@ try {
   return;
 }
 
+// Ensure we have enough memory available for future allocations to succeed.
+if (!binding.ensureAllocation(2 * kStringMaxLength)) {
+  console.log(skipMessage);
+  return;
+}
+
 assert.throws(function() {
-  buf.toString('binary');
+  buf.toString('ascii');
 }, /toString failed/);
-
-var maxString = buf.toString('binary', 1);
-assert.equal(maxString.length, kStringMaxLength);
-// Free the memory early instead of at the end of the next assignment
-maxString = undefined;
-
-maxString = buf.toString('binary', 0, kStringMaxLength);
-assert.equal(maxString.length, kStringMaxLength);
