+++ test/addons/stringbytes-external-exceed-max/test-stringbytes-external-exceed-max.js
@@ -1,7 +1,7 @@
 'use strict';
-// Flags: --expose-gc
 
-const common = require('../common');
+const common = require('../../common');
+const binding = require('./build/Release/binding');
 const assert = require('assert');
 
 const skipMessage =
@@ -10,17 +10,13 @@ if (!common.enoughTestMem) {
   console.log(skipMessage);
   return;
 }
-assert(typeof gc === 'function', 'Run this test with --expose-gc');
 
 // v8 fails silently if string length > v8::String::kMaxLength
 // v8::String::kMaxLength defined in v8.h
 const kStringMaxLength = process.binding('buffer').kStringMaxLength;
 
 try {
-  var buf = new Buffer(kStringMaxLength + 2);
-  // Try to allocate memory first then force gc so future allocations succeed.
-  new Buffer(2 * kStringMaxLength);
-  gc();
+  var buf = new Buffer(kStringMaxLength * 2 + 2);
 } catch (e) {
   // If the exception is not due to memory confinement then rethrow it.
   if (e.message !== 'Invalid array buffer length') throw (e);
@@ -28,5 +24,12 @@ try {
   return;
 }
 
-const maxString = buf.toString('utf16le');
-assert.equal(maxString.length, (kStringMaxLength + 2) / 2);
+// Ensure we have enough memory available for future allocations to succeed.
+if (!binding.ensureAllocation(2 * kStringMaxLength)) {
+  console.log(skipMessage);
+  return;
+}
+
+assert.throws(function() {
+  buf.toString('utf16le');
+}, /toString failed/);
